+++ sfAudio/src/SoundFileReaderWav.cpp
@@ -352,7 +352,7 @@ bool SoundFileReaderWav::parseHeader(Info& info)
             }
 
             // Skip potential extra information
-            if (m_stream->seek(subChunkStart + subChunkSize) == -1)
+            if (m_stream->seek(subChunkStart + subChunkSize) == static_cast<uint64_t>(-1))
                 return false;
         }
         else if ((subChunkId[0] == 'd') && (subChunkId[1] == 'a') && (subChunkId[2] == 't') && (subChunkId[3] == 'a'))
@@ -371,7 +371,7 @@ bool SoundFileReaderWav::parseHeader(Info& info)
         else
         {
             // unknown chunk, skip it
-            if (m_stream->seek(m_stream->tell() + subChunkSize) == -1)
+            if (m_stream->seek(m_stream->tell() + subChunkSize) == static_cast<uint64_t>(-1))
                 return false;
         }
     }
